+++ R/psis.R
@@ -3,7 +3,7 @@
 #' Implementation of Pareto smoothed importance sampling (PSIS), a method for
 #' stabilizing importance ratios. The version of PSIS implemented here
 #' corresponds to the algorithm presented in Vehtari, Simpson, Gelman, Yao,
-#' and Gabry (2019).
+#' and Gabry (2022).
 #' For PSIS diagnostics see the [pareto-k-diagnostic] page.
 #'
 #' @export
@@ -19,11 +19,11 @@
 #'   This is related to the relative efficiency of estimating the normalizing
 #'   term in self-normalizing importance sampling. If `r_eff` is not
 #'   provided then the reported PSIS effective sample sizes and Monte Carlo
-#'   error estimates will be over-optimistic. See the [relative_eff()]
-#'   helper function for computing `r_eff`. If using `psis` with
-#'   draws of the `log_ratios` not obtained from MCMC then the warning
-#'   message thrown when not specifying `r_eff` can be disabled by
-#'   setting `r_eff` to `NA`.
+#'   error estimates can be over-optimistic. If the posterior draws are (near)
+#'   independent then `r_eff=1` can be used. `r_eff` has to be a scalar (same
+#'   value is used for all observations) or a vector with length equal to the
+#'   number of observations. The default value is 1. See the [relative_eff()]
+#'   helper function for computing `r_eff`.
 #'
 #' @return The `psis()` methods return an object of class `"psis"`,
 #'   which is a named list with the following components:
@@ -99,7 +99,7 @@ psis <- function(log_ratios, ...) UseMethod("psis")
 #'
 psis.array <-
   function(log_ratios, ...,
-           r_eff = NULL,
+           r_eff = 1,
            cores = getOption("mc.cores", 1)) {
   importance_sampling.array(log_ratios = log_ratios, ...,
                             r_eff = r_eff,
@@ -115,7 +115,7 @@ psis.array <-
 psis.matrix <-
   function(log_ratios,
            ...,
-           r_eff = NULL,
+           r_eff = 1,
            cores = getOption("mc.cores", 1)) {
     importance_sampling.matrix(log_ratios,
                                ...,
@@ -129,7 +129,7 @@ psis.matrix <-
 #' @template vector
 #'
 psis.default <-
-  function(log_ratios, ..., r_eff = NULL) {
+  function(log_ratios, ..., r_eff = 1) {
     importance_sampling.default(log_ratios = log_ratios, ...,
                                 r_eff = r_eff,
                                 method = "psis")
@@ -273,11 +273,15 @@ psis_smooth_tail <- function(x, cutoff) {
 #' 20% of the total number of weights.
 #'
 #' @noRd
-#' @param r_eff A N-vector of relative MCMC effective sample sizes of `exp(log-lik matrix)`.
+#' @param r_eff A N-vector or scalar of relative MCMC effective sample sizes of
+#'   `exp(log-lik matrix)`. The default value is 1.
 #' @param S The (integer) size of posterior sample.
 #' @return An N-vector of tail lengths.
 #'
 n_pareto <- function(r_eff, S) {
+  if (isTRUE(is.null(r_eff) || all(is.na(r_eff)))) {
+    r_eff <- 1
+  }
   ceiling(pmin(0.2 * S, 3 * sqrt(S / r_eff)))
 }
 
@@ -293,19 +297,16 @@ enough_tail_samples <- function(tail_len, min_len = 5) {
 }
 
 
-#' Throw warnings about pareto k estimates
+#' Throw warnings about Pareto k estimates
 #'
 #' @noRd
-#' @param k A vector of pareto k estimates.
-#' @param high The value at which to warn about slighly high estimates.
-#' @param too_high The value at which to warn about very high estimates.
+#' @param k A vector of Pareto k estimates.
+#' @param k_threshold The value at which to warn about high Pareto k estimates.
 #' @return Nothing, just possibly throws warnings.
 #'
-throw_pareto_warnings <- function(k, high = 0.5, too_high = 0.7) {
-  if (isTRUE(any(k > too_high))) {
+throw_pareto_warnings <- function(k, k_threshold) {
+  if (isTRUE(any(k > k_threshold))) {
     .warn("Some Pareto k diagnostic values are too high. ", .k_help())
-  } else if (isTRUE(any(k > high))) {
-    .warn("Some Pareto k diagnostic values are slightly high. ", .k_help())
   }
 }
 
@@ -347,19 +348,19 @@ throw_tail_length_warnings <- function(tail_lengths) {
 #' @param len The length `r_eff` should have if not `NULL` or `NA`.
 #' @return
 #' * If `r_eff` has length `len` then `r_eff` is returned.
-#' * If `r_eff` is `NULL` then a warning is thrown and `rep(1, len)` is returned.
-#' * If `r_eff` is `NA` then the warning is skipped and
-#'   `rep(1, len)` is returned.
+#' * If `r_eff` is `NULL` then `rep(1, len)` is returned.
+#' * If `r_eff` is `NA` then `rep(1, len)` is returned.
+#' * If `r_eff` is a scalar then `rep(r_eff, len)` is returned.
+#' * If `r_eff` is not a scalar but the length is not `len` then an error is thrown.
 #' * If `r_eff` has length `len` but has `NA`s then an error is thrown.
 #'
 prepare_psis_r_eff <- function(r_eff, len) {
   if (isTRUE(is.null(r_eff) || all(is.na(r_eff)))) {
-    if (!called_from_loo() && is.null(r_eff)) {
-      throw_psis_r_eff_warning()
-    }
     r_eff <- rep(1, len)
+  } else if (length(r_eff) == 1) {
+    r_eff <- rep(r_eff, len)
   } else if (length(r_eff) != len) {
-    stop("'r_eff' must have one value per observation.", call. = FALSE)
+    stop("'r_eff' must have one value or one value per observation.", call. = FALSE)
   } else if (anyNA(r_eff)) {
     stop("Can't mix NA and not NA values in 'r_eff'.", call. = FALSE)
   }
@@ -385,7 +386,7 @@ called_from_loo <- function() {
 throw_psis_r_eff_warning <- function() {
   warning(
     "Relative effective sample sizes ('r_eff' argument) not specified. ",
-    "PSIS n_eff will not be adjusted based on MCMC n_eff.",
+    "PSIS ESS (n_eff) will not be adjusted based on MCMC ESS (n_eff).",
     call. = FALSE
   )
 }